Nevmem/better profile page (#92)
* Loader for profile page

* Checkpoint for feedback

* Fix smth

* Repository

* Feedback manager

* kek

* Network impl

* Some recycler items

* Simple shit

* Fix shit

* Simple feedback setup done

* Add some impls into rating module

* more

* And last

* Implement ratings manage + retry util func

* setup

* Last preparations

* implementation

* Ratings controller

* Fix

+++ android/rating/src/main/java/com/nevmem/qms/rating/Rating.kt
@@ -1,3 +1,3 @@
 package com.nevmem.qms.rating
 
-data class Rating(val score: Double?, val maxScore: Double)
+data class Rating(val score: Float?, val maxScore: Float)
